+++ game/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from character.models.abilities import AbilityType
+from character.constants.abilities import AbilityName
 from character.models.character import Character
 
 from .models.events import RollRequest
@@ -22,11 +22,9 @@ class AbilityCheckRequestForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         game = self.initial["game"]
-
         self.fields["character"] = forms.ModelChoiceField(
             queryset=Character.objects.filter(player__game=game),
             widget=forms.Select(attrs={"class": "rpgui-dropdown"}),
         )
-
         self.fields["ability_type"].label = "Ability"
-        self.fields["ability_type"].choices = AbilityType.AbilityName.choices
+        self.fields["ability_type"].choices = AbilityName.choices